Kmeans running, might return wrong results

+++ IR-Exercise4/src/ClusteringAlgorithms/ICentroid.java
@@ -6,6 +6,6 @@ package ClusteringAlgorithms;
 public interface ICentroid
 {
     double[] GetCoordinates();
-    double GetDistance(double[] otherVector);
+    double GetDistance(float[] otherVector);
     boolean equals(ICentroid other);
 }
